Extract create_macro_block() from NodeService
This is pure refactoring. No functional changes intended.

Needed for new MacroBlocks.
